import() and unimport()
That is a breaking change. We move from using BEGIN and END blocks to
enable and disable Carp::Always to calls of import() and unimport().

This change is not radical, and only affects the following expectations:

* Carp::Always will be enabled at load, as with

    use Carp::Always ()
    require Carp::Always;

* Carp::Always will be disabled automatically at the end
  or the program (which happened with the use of the END block)

The documentation always mentioned

    use Carp::Always

as the usage of the module, which keeps working just fine.
As well as

    use Carp::Always 0.10

Both forms invoke Carp::Always::import().

The introduction of unimport() makes possible to toogle on
and off Carp::Always with sequences of statements like

    use Carp::Always;
    ...    # Carp::Always in effect
    no Carp::Always;
    ...    # no Carp Always anymore