Restore compatibility with Carp < 1.25
Carp 1.25 added a dot to end of produced messages,
to match the formatting from CORE::die

This change makes a tiny change for Carp::Always to
work with Carp < 1.25 as well. In order to do that,
it chomps a dot at the end of the original arguments
(for the case when warn / die have put it there).
This makes the suppression of duplicate tracebacks
to work for this old Carp modules.